Add test for PYSAT with real FMs in CNF
