+++ models/textual_cnf/GPL.txt
@@ -1,0 +1,1 @@
+(GPL) && (!Driver || GPL) && (!GraphType || GPL) && (!Weight || GPL) && (!Search || GPL) && (!Algorithms || GPL) && (!GPL || Driver) && (!GPL || GraphType) && (!GPL || Algorithms) && (!Benchmark || Driver) && (!Driver || Benchmark) && (!Num || Algorithms) && (!CC || Algorithms) && (!Prim || Algorithms) && (!SCC || Algorithms) && (!Kruskal || Algorithms) && (!Cycle || Algorithms) && (!Shortest || Algorithms) && (!Algorithms || CC || Prim || Num || Cycle || SCC || Kruskal || Shortest) && (!Directed || GraphType) && (!Undirected || GraphType) && (!DFS || Search) && (!BFS || Search) && (!Num || Search) && (!CC || Undirected) && (!CC || Search) && (!SCC || Directed) && (!Prim || Undirected) && (!SCC || DFS) && (!Cycle || DFS) && (!Kruskal || Undirected) && (!Kruskal || Weight) && (!Prim || Weight) && (!Kruskal || !Prim) && (!Shortest || Directed) && (!Shortest || Weight)
