+++ models/textual_cnf/aafms_framework-namesAdapted.txt
@@ -1,0 +1,1 @@
+(AAFMFramework) && (!pythonsat || APISolvers) && (!pycosat || APISolvers) && (!pylgl || APISolvers) && (!pyPicosat || APISolvers) && (!pyglucose || APISolvers) && (!pycryptosat || APISolvers) && (!satyrn || APISolvers) && (!satispy || APISolvers) && (!umo || APISolvers) && (!pydepqbf || APISolvers) && (!APISolvers || satyrn || pycryptosat || satispy || pyglucose || pylgl || pycosat || umo || pydepqbf || pythonsat || pyPicosat) && (!DepMng || Packages) && (!APISolvers || Packages) && (!ParserTechnology || Packages) && (!Utils || Packages) && (!Packages || DepMng) && (!pipdeptree || pip) && (!ebnf || larkparser) && (!MiniSAT || Solvers) && (!PicoSAT || Solvers) && (!Glucose || Solvers) && (!CaDiCaL || Solvers) && (!Minicard || Solvers) && (!Maple || Solvers) && (!CryptoMiniSat || Solvers) && (!Lingeling || Solvers) && (!CBC || Solvers) && (!DepQBF || Solvers) && (!Solvers || Glucose || Lingeling || CryptoMiniSat || PicoSAT || CaDiCaL || MiniSAT || CBC || DepQBF || Minicard || Maple) && (!Solvers || AutomatedReasoning) && (!Operations || AutomatedReasoning) && (!AutomatedReasoning || Solvers) && (!ValidModel || Operations) && (!ValidConfiguration || Operations) && (!ValidProduct || Operations) && (!AllProducts || Operations) && (!DeadFeatures || Operations) && (!CoreFeatures || Operations) && (!ErrorDetection || Operations) && (!ErrorDiagnosis || Operations) && (!Operations || ErrorDiagnosis || ErrorDetection || ValidConfiguration || ValidModel || DeadFeatures || ValidProduct || AllProducts || CoreFeatures) && (!Packages || AAFMFramework) && (!Metamodels || AAFMFramework) && (!AutomatedReasoning || AAFMFramework) && (!Interoperability || AAFMFramework) && (!System || AAFMFramework) && (!AAFMFramework || Metamodels) && (!AAFMFramework || System) && (!pip || DepMng) && (!setuptools || DepMng) && (!wheel || DepMng) && (!pkgresources || DepMng) && (!DepMng || pip) && (!DepMng || setuptools) && (!FeatureIDE || Interoperability) && (!SPLOT || Interoperability) && (!UVL || Interoperability) && (!Interoperability || SPLOT || FeatureIDE || UVL) && (!Linux || System) && (!Win || System) && (!System || Linux || Win) && (!Win || !Linux) && (!antlr4python3runtime || ParserTechnology) && (!SimpleParse || ParserTechnology) && (!larkparser || ParserTechnology) && (!ParserTechnology || larkparser || antlr4python3runtime || SimpleParse) && (!FeatureModel || Metamodels) && (!CNFModel || Metamodels) && (!DecisionModel || Metamodels) && (!six || Utils) && (!pybind11 || Utils) && (!Utils || six || pybind11) && (!MiniSAT || satispy || umo || pythonsat) && (!Lingeling || satispy || pylgl || pythonsat) && (!CBC || umo) && (!PicoSAT || pycosat || pyPicosat) && (!Glucose || pythonsat || pyglucose) && (!DepQBF || pydepqbf) && (pythonsat || !CaDiCaL) && (pythonsat || !Maple) && (pythonsat || !Minicard) && (!CryptoMiniSat || pycryptosat) && (pybind11 || !satyrn) && (pybind11 || !pyglucose) && (!pythonsat || six) && (!Win || !umo) && (!Win || !pyPicosat) && (!CNFModel || AutomatedReasoning) && (!AutomatedReasoning || CNFModel) && (!Linux || !pip || pkgresources) && (!UVL || ParserTechnology)
