+++ models/textual_cnf/aafms_framework.txt
@@ -1,0 +1,1 @@
+(AAFMFramework) && (!python-sat || APISolvers) && (!pycosat || APISolvers) && (!pylgl || APISolvers) && (!pyPicosat || APISolvers) && (!pyglucose || APISolvers) && (!pycryptosat || APISolvers) && (!satyrn || APISolvers) && (!satispy || APISolvers) && (!umo || APISolvers) && (!pydepqbf || APISolvers) && (!APISolvers || satyrn || pycryptosat || satispy || pyglucose || pylgl || pycosat || umo || python-sat || pydepqbf || pyPicosat) && (!DepMng || Packages) && (!APISolvers || Packages) && (!ParserTechnology || Packages) && (!Utils || Packages) && (!Packages || DepMng) && (!pipdeptree || pip) && (!MiniSAT || Solvers) && (!PicoSAT || Solvers) && (!Glucose || Solvers) && (!CaDiCaL || Solvers) && (!Minicard || Solvers) && (!Maple || Solvers) && (!CryptoMiniSat || Solvers) && (!Lingeling || Solvers) && (!CBC || Solvers) && (!DepQBF || Solvers) && (!Solvers || Glucose || Lingeling || CryptoMiniSat || PicoSAT || CaDiCaL || MiniSAT || CBC || DepQBF || Minicard || Maple) && (!Solvers || AutomatedReasoning) && (!Operations || AutomatedReasoning) && (!AutomatedReasoning || Solvers) && (!ValidModel || Operations) && (!ValidConfiguration || Operations) && (!ValidProduct || Operations) && (!AllProducts || Operations) && (!DeadFeatures || Operations) && (!CoreFeatures || Operations) && (!ErrorDetection || Operations) && (!ErrorDiagnosis || Operations) && (!Operations || ErrorDiagnosis || ErrorDetection || ValidConfiguration || ValidModel || DeadFeatures || ValidProduct || AllProducts || CoreFeatures) && (!Packages || AAFMFramework) && (!Metamodels || AAFMFramework) && (!AutomatedReasoning || AAFMFramework) && (!Interoperability || AAFMFramework) && (!System || AAFMFramework) && (!AAFMFramework || Metamodels) && (!AAFMFramework || System) && (!pip || DepMng) && (!setuptools || DepMng) && (!wheel || DepMng) && (!pkg-resources || DepMng) && (!DepMng || pip) && (!DepMng || setuptools) && (!FeatureIDE || Interoperability) && (!SPLOT || Interoperability) && (!UVL || Interoperability) && (!Interoperability || SPLOT || FeatureIDE || UVL) && (!Linux || System) && (!Win || System) && (!System || Linux || Win) && (!Win || !Linux) && (!antlr4-python3-runtime || ParserTechnology) && (!SimpleParse || ParserTechnology) && (!lark-parser || ParserTechnology) && (!ParserTechnology || antlr4-python3-runtime || lark-parser || SimpleParse) && (!ebnf || lark-parser) && (!FeatureModel || Metamodels) && (!CNFModel || Metamodels) && (!DecisionModel || Metamodels) && (!six || Utils) && (!pybind11 || Utils) && (!Utils || six || pybind11) && (!MiniSAT || satispy || umo || python-sat) && (!Lingeling || satispy || pylgl || python-sat) && (!CBC || umo) && (!PicoSAT || pycosat || pyPicosat) && (!Glucose || python-sat || pyglucose) && (!DepQBF || pydepqbf) && (python-sat || !CaDiCaL) && (python-sat || !Maple) && (python-sat || !Minicard) && (!CryptoMiniSat || pycryptosat) && (pybind11 || !satyrn) && (pybind11 || !pyglucose) && (!python-sat || six) && (!Win || !umo) && (!Win || !pyPicosat) && (!CNFModel || AutomatedReasoning) && (!AutomatedReasoning || CNFModel) && (!Linux || !pip || pkg-resources) && (!UVL || ParserTechnology)
