+++ models/textual_cnf/jHipster.txt
@@ -1,0 +1,1 @@
+(JHipster) && (!HazelCast || Hibernate2ndLvlCache) && (!EhCache || Hibernate2ndLvlCache) && (!Hibernate2ndLvlCache || EhCache || HazelCast) && (!HazelCast || !EhCache) && (!MicroserviceApplication || Server) && (!UaaServer || Server) && (!Server || MicroserviceApplication || UaaServer) && (!UaaServer || !MicroserviceApplication) && (!Gradle || BackEnd) && (!Maven || BackEnd) && (!BackEnd || Maven || Gradle) && (!Gradle || !Maven) && (!Generator || JHipster) && (!Authentication || JHipster) && (!SocialLogin || JHipster) && (!Database || JHipster) && (!SpringWebSockets || JHipster) && (!Libsass || JHipster) && (!ClusteredSession || JHipster) && (!BackEnd || JHipster) && (!InternationalizationSupport || JHipster) && (!Docker || JHipster) && (!TestingFrameworks || JHipster) && (!JHipster || Generator) && (!JHipster || Authentication) && (!JHipster || BackEnd) && (!JHipster || TestingFrameworks) && (!MySQL || Production) && (!MariaDB || Production) && (!PostgreSQL || Production) && (!Production || PostgreSQL || MySQL || MariaDB) && (!MariaDB || !MySQL) && (!MySQL || !PostgreSQL) && (!MariaDB || !PostgreSQL) && (!SQL || Database) && (!Cassandra || Database) && (!MongoDB || Database) && (!Database || Cassandra || SQL || MongoDB) && (!SQL || !Cassandra) && (!MongoDB || !SQL) && (!MongoDB || !Cassandra) && (!H2 || Development) && (!PostgreeSQLDev || Development) && (!MariaDBDev || Development) && (!MySql || Development) && (!Development || MySql || PostgreeSQLDev || H2 || MariaDBDev) && (!H2 || !PostgreeSQLDev) && (!MariaDBDev || !H2) && (!H2 || !MySql) && (!MariaDBDev || !PostgreeSQLDev) && (!PostgreeSQLDev || !MySql) && (!MariaDBDev || !MySql) && (!DiskBased || H2) && (!InMemory || H2) && (!H2 || InMemory || DiskBased) && (!DiskBased || !InMemory) && (!Hibernate2ndLvlCache || SQL) && (!Development || SQL) && (!Production || SQL) && (!ElasticSearch || SQL) && (!SQL || Development) && (!SQL || Production) && (!HTTPSession || Authentication) && (!OAuth2 || Authentication) && (!Uaa || Authentication) && (!JWT || Authentication) && (!Authentication || OAuth2 || JWT || HTTPSession || Uaa) && (!HTTPSession || !OAuth2) && (!Uaa || !HTTPSession) && (!HTTPSession || !JWT) && (!Uaa || !OAuth2) && (!JWT || !OAuth2) && (!Uaa || !JWT) && (!Protractor || TestingFrameworks) && (!Gatling || TestingFrameworks) && (!Cucumber || TestingFrameworks) && (!TestingFrameworks || Gatling) && (!TestingFrameworks || Cucumber) && (!MicroserviceGateway || Application) && (!Monolithic || Application) && (!Application || MicroserviceGateway || Monolithic) && (!Monolithic || !MicroserviceGateway) && (!Server || Generator) && (!Application || Generator) && (!Generator || Server || Application) && (!Application || !Server) && (!OAuth2 || SocialLogin || MicroserviceApplication || SQL || MongoDB) && (!SocialLogin || HTTPSession || JWT) && (!SocialLogin || SQL || MongoDB) && (!SocialLogin || Monolithic) && (!UaaServer || Uaa) && (OAuth2 || SocialLogin || MicroserviceApplication || SQL || MongoDB || Cassandra) && (!Server || !Protractor) && (Server || Protractor) && (!MySQL || H2 || MySql) && (JWT || Uaa || !MicroserviceApplication) && (JWT || Uaa || !MicroserviceGateway) && (!Monolithic || JWT || HTTPSession || OAuth2) && (!MariaDB || H2 || MariaDBDev) && (!PostgreSQL || H2 || PostgreeSQLDev) && (Application || !SpringWebSockets) && (Application || !ClusteredSession) && (!Libsass || Application)
