+++ models/textual_cnf/mobilemedia2.txt
@@ -1,0 +1,1 @@
+(MobileMedia2) && (!SetFavourites || Favourites) && (!ViewFavourites || Favourites) && (!Favourites || SetFavourites) && (!Favourites || ViewFavourites) && (!MediaSelection || MobileMedia2) && (!MediaManagement || MobileMedia2) && (!ScreenSize || MobileMedia2) && (!Speakerphone || MobileMedia2) && (!Calendar || MobileMedia2) && (!Radio || MobileMedia2) && (!CallManager || MobileMedia2) && (!TVDigital || MobileMedia2) && (!OS || MobileMedia2) && (!TouchScreen || MobileMedia2) && (!Connectivity || MobileMedia2) && (!GPS || MobileMedia2) && (!ResolutionCamera || MobileMedia2) && (!Language || MobileMedia2) && (!MobileMedia2 || MediaSelection) && (!MobileMedia2 || MediaManagement) && (!MobileMedia2 || ScreenSize) && (!MobileMedia2 || Speakerphone) && (!MobileMedia2 || Calendar) && (!MobileMedia2 || Radio) && (!MobileMedia2 || CallManager) && (!MobileMedia2 || OS) && (!MobileMedia2 || Language) && (!Portuguese || Language) && (!English || Language) && (!Spanish || Language) && (!Language || Portuguese || Spanish || English) && (!Screen1 || ScreenSize) && (!Screen2 || ScreenSize) && (!Screen3 || ScreenSize) && (!ScreenSize || Screen2 || Screen3 || Screen1) && (!Screen1 || !Screen2) && (!Screen1 || !Screen3) && (!Screen3 || !Screen2) && (!Favourites || MediaManagement) && (!CopyMedia || MediaManagement) && (!SMSTransfer || MediaManagement) && (!ReceivePhoto || SMSTransfer) && (!SendPhoto || SMSTransfer) && (!Androide || OS) && (!WindowsMobile || OS) && (!Symbia || OS) && (!iOS || OS) && (!OS || iOS || Symbia || Androide || WindowsMobile) && (!WindowsMobile || !Androide) && (!Androide || !Symbia) && (!Androide || !iOS) && (!WindowsMobile || !Symbia) && (!WindowsMobile || !iOS) && (!Symbia || !iOS) && (!Resolution3MP || ResolutionCamera) && (!Resolution8MP || ResolutionCamera) && (!ResolutionCamera || Resolution3MP || Resolution8MP) && (!Resolution8MP || !Resolution3MP) && (!MusicPlayer || Music) && (!Music || MusicPlayer) && (!Bluetooth || Connectivity) && (!WiFi || Connectivity) && (!Connectivity || Bluetooth || WiFi) && (!Gmail || WiFi) && (!GoogleMaps || WiFi) && (!Youtube || WiFi) && (!WiFi || Youtube || GoogleMaps || Gmail) && (!Photo || MediaSelection) && (!Music || MediaSelection) && (!Video || MediaSelection) && (!MediaSelection || Photo || Music || Video) && (Video || !ResolutionCamera) && (Photo || !ResolutionCamera) && (!CopyMedia || SMSTransfer)
