+++ models/textual_cnf/pizzas.txt
@@ -1,0 +1,1 @@
+(Pizza) && (!Normal || Size) && (!Big || Size) && (!Size || Big || Normal) && (!Normal || !Big) && (!Topping || Pizza) && (!Size || Pizza) && (!Dough || Pizza) && (!CheesyCrust || Pizza) && (!Pizza || Topping) && (!Pizza || Size) && (!Pizza || Dough) && (!Salami || Topping) && (!Ham || Topping) && (!Mozzarella || Topping) && (!Topping || Mozzarella || Salami || Ham) && (!Neapolitan || Dough) && (!Sicilian || Dough) && (!Dough || Neapolitan || Sicilian) && (!Sicilian || !Neapolitan) && (!CheesyCrust || Big)
