+++ models/textual_cnf/tankwar.txt
@@ -1,0 +1,1 @@
+(TankWar) && (!PC || Platform) && (!Handy || Platform) && (!Platform || Handy || PC) && (!PC || !Handy) && (!fuer_PC || Image) && (!fuer_Handy || Image) && (!IMG_tool || Image) && (!EN || Language) && (!DE || Language) && (!Language || DE || EN) && (!EN || !DE) && (!Map || GUI) && (!Image || GUI) && (!GUI || Map) && (!Re_fuer_PC || Record) && (!Re_fuer_Handy || Record) && (!Record || Re_fuer_PC || Re_fuer_Handy) && (!Re_fuer_Handy || !Re_fuer_PC) && (!Sound_fuer_pc || Soundeffekt) && (!Sound_fuer_Handy || Soundeffekt) && (!Soundeffekt || Sound_fuer_pc || Sound_fuer_Handy) && (!Sound_fuer_Handy || !Sound_fuer_pc) && (!USA_M1Abrams || Tanks) && (!Germany_Leopard || Tanks) && (!China_Type99 || Tanks) && (!Tanks || USA_M1Abrams || Germany_Leopard || China_Type99) && (!Beschleunigung || Tools) && (!einfrieren || Tools) && (!Bombe || Tools) && (!Energie || Tools) && (!Feuerkraft || Tools) && (!Mars || Tools) && (!Tools || Beschleunigung || Energie || Feuerkraft || Mars || einfrieren || Bombe) && (!Easy || AI) && (!Hard || AI) && (!AI || Easy || Hard) && (!Hard || !Easy) && (!Platform || TankWar) && (!Tools || TankWar) && (!explodieren || TankWar) && (!GUI || TankWar) && (!Record || TankWar) && (!Soundeffekt || TankWar) && (!Language || TankWar) && (!Tanks || TankWar) && (!AI || TankWar) && (!TankWar || Platform) && (!TankWar || GUI) && (!TankWar || Language) && (!TankWar || Tanks) && (!TankWar || AI) && (!M_240 || Map) && (!M_600 || Map) && (!M_780 || Map) && (!Map || M_780 || M_240 || M_600) && (!M_600 || !M_240) && (!M_240 || !M_780) && (!M_600 || !M_780)
