+++ models/textual_cnf/wget.txt
@@ -1,0 +1,1 @@
+(wget) && (!base || wget) && (!output_file || wget) && (!server_response || wget) && (!no_clobber || wget) && (!no_dns_cache || wget) && (!verbosity || wget) && (!no_directories || wget) && (!no_host_directories || wget) && (!no_http_keep_alive || wget) && (!no_cookies || wget) && (!save_headers || wget) && (!convert_links || wget) && (!strict_comments || wget) && (!wget || base) && (!no_verbose || verbosity) && (!verbose || verbosity) && (!quiet || verbosity) && (!verbosity || verbose || no_verbose || quiet) && (!no_verbose || !verbose) && (!quiet || !no_verbose) && (!quiet || !verbose)
