+++ operations/fm_metamodel/FMMetricsTest.py
@@ -6,7 +6,7 @@ from models.models_info import *
 
 
 def get_model(model_name) -> str:
-    return FeatureIDEReader(INPUT_MODELS_FOLDER + model_name + EXTENSION).transform()
+    return FeatureIDEReader(INPUT_FIDE_MODELS_FOLDER + model_name + FIDE_EXTENSION).transform()
 
 
 @pytest.mark.parametrize("model_name, expected_nof_features", [[m[NAME], m[NOF_FEATURES]] for m in MODELS])
